requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.9.0
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.2
greenletio==0.9.0
gunicorn==20.0.4
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.4
pycryptodomex==3.17
python-dotenv==0.21.1
regex==2022.10.31
requests==2.28.2
revChatGPT==1.1.3
tiktoken==0.2.0
tqdm==4.64.1
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.12.0
